+++ src/Resources/contao/languages/de/tl_module.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * Translations are managed using Transifex. To create a new translation
+ * or to help to maintain an existing one, please register at transifex.com.
+ *
+ * @link https://www.transifex.com/signup/
+ * @link https://www.transifex.com/projects/p/contao-bootstrap/language/de/
+ *
+ * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
+ *
+ * last-updated: 2017-09-28T10:33:15+00:00
+ */
+
+
+$GLOBALS['TL_LANG']['tl_module']['bs_grid']['0']                 = 'Grid';
+$GLOBALS['TL_LANG']['tl_module']['bs_grid']['1']                 = 'Bitte wählen Sie ein vordefiniertes Grid aus.';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModule']['0']           = 'Modul oder Spalte';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModule']['1']           = 'Wählen Sie ein Modul aus oder legen Sie die nächste Spalte fest.';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModule']['grid']        = 'Grid';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModule']['module']      = 'Modul';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModules']['0']          = 'Module';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModules']['1']          = 'Wählen Sie die Module und Spalten aus die Sie kombinieren möchten.';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModules_inactive']['0'] = 'Inaktiv';
+$GLOBALS['TL_LANG']['tl_module']['bs_gridModules_inactive']['1'] = 'Deaktivieren Sie diesen Eintrag zeitweise.';
+$GLOBALS['TL_LANG']['tl_module']['bs_separatorTitle']            = '----- Nächste Spalte -----';
+